+++ pkg/dds/lib/src/dap/protocol_converter.dart
@@ -609,8 +609,7 @@ class ProtocolConverter {
     final uriIsPackage = uri?.isScheme('package') ?? false;
     final sourcePathUri =
         uri != null ? await thread.resolveUriToPath(uri) : null;
-    var canShowSource =
-        sourcePathUri != null && _adapter.isSupportedFileScheme(sourcePathUri);
+    var canShowSource = sourcePathUri != null && sourcePathUri.isScheme('file');
 
     // If we don't have a local source file but the source is a "dart:" uri we
     // might still be able to download the source from the VM.
@@ -639,7 +638,7 @@ class ProtocolConverter {
     }
 
     // LSP uses 0 for unknown lines.
-    var (line, col) = lineCol ?? (0, 0);
+    final (line, col) = lineCol ?? (0, 0);
 
     // If a source would be considered not-debuggable (for example it's in the
     // SDK and debugSdkLibraries=false) then we should also mark it as