+++ pkg/dds/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 5.2.0-wip
 
 - [DAP] `Stopped(reason: 'entry')` events will no longer be lost if an isolate has not yet reached the `PauseStart` state when connecting to the VM.
+- **Breaking change:** [DAP] Support for the custom `supportsDartUris` client capability and `dart-macro+file:///` mappings that supported the Dart macros experiment have been removed.
 
 # 5.1.0
 - Update to version 2.1 of the DDS protocol.