+++ pkg/dds/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 5.2.0-wip
+# 5.2.0
 
 - [DAP] `Stopped(reason: 'entry')` events will no longer be lost if an isolate has not yet reached the `PauseStart` state when connecting to the VM.
 - **Breaking change:** [DAP] Support for the custom `supportsDartUris` client capability and `dart-macro+file:///` mappings that supported the Dart macros experiment have been removed.